DVD-rental/dvdrental-Mohamed_Ashraf/requirements.txt
djangorestframework
django
# REQUIRED
pip install django
pip install djangorestframework
# Postgres
pip install pillow
pip install psycopg2